requirements.txt
openai-whisper
sounddevice
soundfile
numpy
pynput
pyperclip
colorama
torch

